Move to samples / general
