+++ docs/_views.md
@@ -6,5 +6,5 @@
 | [`scores_timeline()`](view-scores-timeline.qmd) | Scatter plot with eval scores by model, organization, and release date. Filterable by evaluation and organization. |
 | [`scores_heatmap()`](view-scores-heatmap.qmd) |  Heatmap with values for comparing scores across model and task. |
 | [`scores_by_model()`](view-scores-by-model.qmd) | Bar plot for comparing model scores on a single eval. |
-| [`tool_calls()`](view-tool-calls.qmd) | Heat map visualising tool calls over evaluation turns. |
-
+| [`samples_tool_calls()`](view-samples-tool-calls.qmd) | Heat map visualising tool calls over evaluation turns. |
+| [`samples_heatmap()`](view-samples-heatmap.qmd) | Heat map visualising sample scores. |
