+++ docs/examples/inspect/samples-tool-calls/cybench_tools.parquet
