+++ docs/reference/_sidebar.yml
@@ -44,8 +44,12 @@ website:
           href: reference/inspect_viz.view.qmd#scores_by_limit
         - text: scores_by_limit_df
           href: reference/inspect_viz.view.qmd#scores_by_limit_df
-        - text: tool_calls
-          href: reference/inspect_viz.view.qmd#tool_calls
+        - text: samples_tool_calls
+          href: reference/inspect_viz.view.qmd#samples_tool_calls
+        - text: samples_heatmap
+          href: reference/inspect_viz.view.qmd#samples_heatmap
+        - text: heatmap
+          href: reference/inspect_viz.view.qmd#heatmap
         - text: CellOptions
           href: reference/inspect_viz.view.qmd#celloptions
       - section: inspect_viz.plot
