+++ src/inspect_viz/view/beta/__init__.py
@@ -1,12 +1,12 @@
-from ...plot._heatmap import CellOptions
+from ._heatmap import CellOptions, heatmap
 from ._samples_heatmap import samples_heatmap
+from ._samples_tool_calls import samples_tool_calls
 from ._scores_by_factor import scores_by_factor
 from ._scores_by_limit import scores_by_limit, scores_by_limit_df
 from ._scores_by_model import scores_by_model
 from ._scores_by_task import scores_by_task
 from ._scores_heatmap import scores_heatmap
 from ._scores_timeline import scores_timeline
-from ._tool_calls import tool_calls
 
 __all__ = [
     "scores_by_factor",
@@ -14,10 +14,11 @@ __all__ = [
     "scores_timeline",
     "scores_heatmap",
     "scores_by_model",
-    "tool_calls",
+    "samples_tool_calls",
     "scores_heatmap",
     "samples_heatmap",
     "CellOptions",
     "scores_by_limit_df",
     "scores_by_limit",
+    "heatmap",
 ]
