+++ src/inspect_viz/view/beta/_heatmap.py
