+++ src/inspect_viz/view/beta/_samples_heatmap.py
@@ -8,8 +8,8 @@ from inspect_viz.mark._channel import SortOrder
 from inspect_viz.mark._mark import Marks
 from inspect_viz.mark._title import Title
 from inspect_viz.plot._attributes import PlotAttributes
-from inspect_viz.plot._heatmap import CellOptions, heatmap
 from inspect_viz.plot._legend import Legend
+from inspect_viz.view.beta._heatmap import CellOptions, heatmap
 
 
 def samples_heatmap(
@@ -87,7 +87,7 @@ def samples_heatmap(
         color_scale="linear",
         x_scale="band",
         padding=0,
-        color_scheme="greens",
+        color_scheme="purples",
         color_domain=color_domain,
     )
     attributes = defaultAttributes | attributes
